requirements.txt
colorama==0.4.1
cryptography==2.8
pytest_cov==2.7.1
pynacl==1.3.0
